requirements.txt
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0
requests==2.31.0
psutil==5.9.8